requirements.txt
Django==2.1.7
boto3==1.4.7
celery==4.1.0
elasticsearch==6.2.0
djangorestframework==3.8.2
kombu==4.5.0
SQLAlchemy==1.2.8
gunicorn==19.8.1,
requests>=2.20.0